refactor: rename strategy/ to adapter/ to distinguish from strategies/
