+++ live_trading_bot/adapter/adapter.py
