+++ live_trading_bot/adapter/ensemble.py
